+++ fvm/src/kernel/default.rs
@@ -31,7 +31,7 @@ use crate::call_manager::{CallManager, InvocationResult, NO_DATA_BLOCK_ID};
 use crate::externs::{Consensus, Rand};
 use crate::gas::GasCharge;
 use crate::state_tree::ActorState;
-use crate::{syscall_error, EMPTY_ARR_CID};
+use crate::syscall_error;
 
 lazy_static! {
     static ref NUM_CPUS: usize = num_cpus::get();
@@ -223,10 +223,7 @@ where
             // exists; if missing, it fails the self destruct.
             //
             // In FVM we check unconditionally, since we only support nv13+.
-            let beneficiary_id = self
-                .resolve_address(beneficiary)?
-                .context("beneficiary doesn't exist")
-                .or_error(ErrorNumber::NotFound)?;
+            let beneficiary_id = self.resolve_address(beneficiary)?;
 
             if beneficiary_id == self.actor_id {
                 return Err(syscall_error!(Forbidden, "benefactor cannot be beneficiary").into());
@@ -679,20 +676,19 @@ where
         self.call_manager.context().network_context.timestamp
     }
 
-    fn tipset_cid(&self, epoch: i64) -> Result<Option<Cid>> {
+    fn tipset_cid(&self, epoch: ChainEpoch) -> Result<Cid> {
         if epoch < 0 {
             return Err(syscall_error!(IllegalArgument; "epoch is negative").into());
         }
-        if epoch >= 900 {
-            return Err(syscall_error!(IllegalArgument; "epoch out of finality range").into());
+        let offset = self.call_manager.context().network_context.epoch - epoch;
+        if offset < 0 {
+            return Err(syscall_error!(IllegalArgument; "epoch is in the future").into());
         }
-
         let tipsets = &self.call_manager.context().network_context.tipsets;
-        if (epoch as usize) < tipsets.len() {
-            return Ok(Some(tipsets[epoch as usize]));
+        if offset >= tipsets.len() as i64 {
+            return Err(syscall_error!(LimitExceeded; "tipset lookback exceeded limit").into());
         }
-
-        Ok(None)
+        Ok(tipsets[offset as usize])
     }
 }
 
@@ -745,18 +741,23 @@ impl<C> ActorOps for DefaultKernel<C>
 where
     C: CallManager,
 {
-    fn resolve_address(&self, address: &Address) -> Result<Option<ActorID>> {
-        self.call_manager.state_tree().lookup_id(address)
+    fn resolve_address(&self, address: &Address) -> Result<ActorID> {
+        Ok(self
+            .call_manager
+            .state_tree()
+            .lookup_id(address)?
+            .ok_or_else(|| syscall_error!(NotFound; "actor not found"))?)
     }
 
-    fn get_actor_code_cid(&self, id: ActorID) -> Result<Option<Cid>> {
+    fn get_actor_code_cid(&self, id: ActorID) -> Result<Cid> {
         Ok(self
             .call_manager
             .state_tree()
             .get_actor_id(id)
             .context("failed to lookup actor to get code CID")
             .or_fatal()?
-            .map(|act| act.code))
+            .ok_or_else(|| syscall_error!(NotFound; "actor not found"))?
+            .code)
     }
 
     // TODO(M2) merge new_actor_address and create_actor into a single syscall.
@@ -781,7 +782,12 @@ where
     }
 
     // TODO(M2) merge new_actor_address and create_actor into a single syscall.
-    fn create_actor(&mut self, code_id: Cid, actor_id: ActorID) -> Result<()> {
+    fn create_actor(
+        &mut self,
+        code_id: Cid,
+        actor_id: ActorID,
+        predictable_address: Option<Address>,
+    ) -> Result<()> {
         // TODO https://github.com/filecoin-project/builtin-actors/issues/492
         let singleton = self
             .call_manager
@@ -795,19 +801,48 @@ where
             );
         }
 
-        let state_tree = self.call_manager.state_tree();
-        if let Ok(Some(_)) = state_tree.get_actor_id(actor_id) {
-            return Err(syscall_error!(Forbidden; "Actor address already exists").into());
-        }
+        // Check to make sure the actor doesn't exist, or is an embryo.
+        let actor = match self.call_manager.state_tree().get_actor_id(actor_id)? {
+            // Replace the embryo
+            Some(mut act)
+                if self
+                    .call_manager
+                    .machine()
+                    .builtin_actors()
+                    .is_embryo_actor(&act.code) =>
+            {
+                if act.address.is_none() {
+                    // The FVM made a mistake somewhere.
+                    return Err(ExecutionError::Fatal(anyhow!(
+                        "embryo {actor_id} doesn't have a predictable address"
+                    )));
+                }
+                if act.address != predictable_address {
+                    // The Init actor made a mistake?
+                    return Err(syscall_error!(
+                        Forbidden,
+                        "embryo has a different predictable address"
+                    )
+                    .into());
+                }
+                act.code = code_id;
+                act
+            }
+            // Don't replace anything else.
+            Some(_) => {
+                return Err(syscall_error!(Forbidden; "Actor address already exists").into());
+            }
+            // Create a new actor.
+            None => {
+                self.call_manager
+                    .charge_gas(self.call_manager.price_list().on_create_actor())?;
+                ActorState::new_empty(code_id, predictable_address)
+            }
+        };
 
         self.call_manager
-            .charge_gas(self.call_manager.price_list().on_create_actor())?;
-
-        let state_tree = self.call_manager.state_tree_mut();
-        state_tree.set_actor_id(
-            actor_id,
-            ActorState::new(code_id, *EMPTY_ARR_CID, TokenAmount::zero(), 0),
-        )
+            .state_tree_mut()
+            .set_actor_id(actor_id, actor)
     }
 
     fn get_builtin_actor_type(&self, code_cid: &Cid) -> u32 {
@@ -846,6 +881,15 @@ where
             .map(|a| a.balance)
             .unwrap_or_default();
         Ok(balance)
+    }
+
+    fn lookup_address(&self, actor_id: ActorID) -> Result<Option<Address>> {
+        Ok(self
+            .call_manager
+            .state_tree()
+            .get_actor_id(actor_id)?
+            .ok_or_else(|| syscall_error!(NotFound; "actor not found"))?
+            .address)
     }
 }