+++ sdk/src/actor.rs
@@ -1,10 +1,12 @@
-use core::option::Option; // no_std
+use core::option::Option;
+use std::ptr; // no_std
 
 use cid::Cid;
 use fvm_shared::address::{Address, Payload};
 use fvm_shared::econ::TokenAmount;
 use fvm_shared::error::ErrorNumber;
 use fvm_shared::{ActorID, MAX_CID_LEN};
+use log::error;
 
 use crate::{sys, SyscallResult, MAX_ACTOR_ADDR_LEN};
 
@@ -20,6 +22,36 @@ pub fn resolve_address(addr: &Address) -> Option<ActorID> {
     unsafe {
         match sys::actor::resolve_address(bytes.as_ptr(), bytes.len() as u32) {
             Ok(value) => Some(value),
+            Err(ErrorNumber::NotFound) => None,
+            Err(other) => panic!("unexpected address resolution failure: {}", other),
+        }
+    }
+}
+
+/// Looks up the f1, f3, or f4 address of the specified actor. Returns `None` if the actor doesn't
+/// exist or it doesn't have an f1, f3, or f4 address.
+pub fn lookup_address(addr: ActorID) -> Option<Address> {
+    let mut out_buffer = [0u8; MAX_ACTOR_ADDR_LEN];
+    unsafe {
+        match sys::actor::lookup_address(addr, out_buffer.as_mut_ptr(), out_buffer.len() as u32) {
+            Ok(0) => None,
+            Ok(length) => match Address::from_bytes(&out_buffer[..length as usize]) {
+                Ok(addr) => Some(addr),
+                // Ok, so, we _log_ this error (if debugging is enabled) but otherwise move on.
+                // Why? Because the system may add _new_ address classes. In that case, the "least
+                // bad" thing to do here is to claim that the target actor doesn't have an f1/f3/f4
+                // address, which is likely correct.
+                //
+                // https://github.com/filecoin-project/builtin-actors/issues/738
+                Err(e) => {
+                    error!(
+                        "unexpected address from 'lookup_address' with protocol {}: {}",
+                        out_buffer[0], e
+                    );
+                    None
+                }
+            },
+            // We're flattening the "not found" error here, but that's probably reasonable for most users.
             Err(ErrorNumber::NotFound) => None,
             Err(other) => panic!("unexpected address resolution failure: {}", other),
         }
@@ -42,8 +74,7 @@ pub fn get_actor_code_cid(addr: &Address) -> Option<Cid> {
     }
 }
 
-/// Generates a new actor address for an actor deployed
-/// by the calling actor.
+/// Generates a new actor address for an actor deployed by the calling actor.
 pub fn new_actor_address() -> Address {
     let mut buf = [0u8; MAX_ACTOR_ADDR_LEN];
     unsafe {
@@ -53,12 +84,21 @@ pub fn new_actor_address() -> Address {
     }
 }
 
-/// Creates a new actor of the specified type in the state tree, under
-/// the provided address.
-/// TODO(M2): this syscall will change to calculate the address internally.
-pub fn create_actor(actor_id: ActorID, code_cid: &Cid) -> SyscallResult<()> {
-    let cid = code_cid.to_bytes();
-    unsafe { sys::actor::create_actor(actor_id, cid.as_ptr()) }
+/// Creates a new actor of the specified type in the state tree, under the provided address.
+pub fn create_actor(
+    actor_id: ActorID,
+    code_cid: &Cid,
+    predictable_address: Option<Address>,
+) -> SyscallResult<()> {
+    unsafe {
+        let cid = code_cid.to_bytes();
+        let addr_bytes = predictable_address.map(|addr| addr.to_bytes());
+        let (addr_off, addr_len) = addr_bytes
+            .as_deref()
+            .map(|v| (v.as_ptr(), v.len()))
+            .unwrap_or((ptr::null(), 0));
+        sys::actor::create_actor(actor_id, cid.as_ptr(), addr_off, addr_len as u32)
+    }
 }
 
 /// Installs or ensures an actor code CID is valid and loaded.
@@ -97,11 +137,13 @@ pub fn get_code_cid_for_type(typ: i32) -> Cid {
     }
 }
 
-/// Retrieves the balance associated with an actor
-pub fn balance_of(actor_id: ActorID) -> TokenAmount {
+/// Retrieves the balance of the specified actor, or None if the actor doesn't exist.
+pub fn balance_of(actor_id: ActorID) -> Option<TokenAmount> {
     unsafe {
-        sys::actor::balance_of(actor_id)
-            .expect("failed to get actor balance")
-            .into()
+        match sys::actor::balance_of(actor_id) {
+            Ok(balance) => Some(balance.into()),
+            Err(ErrorNumber::NotFound) => None,
+            Err(e) => panic!("unexpected error: {e}"),
+        }
     }
 }