+++ sdk/src/sys/network.rs
@@ -1,5 +1,9 @@
 //! Syscalls for network metadata.
 
+// for documentation links
+#[cfg(doc)]
+use crate::sys::ErrorNumber::*;
+
 super::fvm_syscalls! {
     module = "network";
 
@@ -26,9 +30,22 @@ super::fvm_syscalls! {
 
     /// Retrieves a tipset's CID within the last finality, if available
     ///
+    /// # Arguments
+    ///
+    /// - `epoch` the epoch being queried.
+    /// - `ret_off` and `ret_len` specify the location and length of the buffer into which the
+    ///   tipset CID will be written.
+    ///
+    /// # Returns
+    ///
+    /// Returns the length of the CID written to the output buffer.
+    ///
     /// # Errors
     ///
-    /// IllegalArgument -- raised when the epoch is negative or greater/equal than finality.
+    /// | Error               | Reason                                       |
+    /// |---------------------|----------------------------------------------|
+    /// | [`IllegalArgument`] | specified epoch is negative or in the future |
+    /// | [`LimitExceeded`]   | specified epoch exceeds finality             |
     pub fn tipset_cid(
         epoch: i64,
         ret_off: *mut u8,