+++ sdk/src/vm.rs
@@ -1,5 +1,6 @@
 use std::ptr;
 
+use fvm_shared::error::ExitCode;
 use fvm_shared::sys::out::vm::InvocationContext;
 
 use crate::sys;
@@ -26,4 +27,17 @@ pub fn abort(code: u32, message: Option<&str>) -> ! {
 
         sys::vm::abort(code, message, message_len as u32);
     }
+}
+
+/// Sets a panic handler to turn all panics into aborts with `USR_ASSERTION_FAILED`. This should be
+/// called early in the actor to improve debuggability.
+///
+/// NOTE: This will incure a small cost on failure (to format an error message).
+pub fn set_panic_handler() {
+    std::panic::set_hook(Box::new(|info| {
+        abort(
+            ExitCode::USR_ASSERTION_FAILED.value(),
+            Some(&format!("{}", info)),
+        )
+    }));
 }