+++ ipld/bitfield/CHANGELOG.md
@@ -4,6 +4,10 @@ Changes to Filecoin's Bitfield library.
 
 ## [Unreleased]
 
+## 0.5.4 [2022-10-11]
+
+- Bumps `fvm_ipld_encoding` and switches from `cs_serde_bytes` to `fvm_ipld_encoding::strict_bytes`.
+
 ## 0.5.3 [2022-09-12]
 
 - Optimize no-op operations.