+++ ipld/hamt/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+## 0.6.0
+
+- Bumps `fvm_ipld_encoding` and switches from `cs_serde_bytes` to `fvm_ipld_encoding::strict_bytes`.
+
 ## 0.5.1
 
 - Update `fvm_ipld_encoding` to 0.2.0.