+++ shared/CHANGELOG.md
@@ -1,6 +1,23 @@
 # Changelog
 
-## 3.0.0-alpha.3 [UNRELEASED]
+
+## 3.0.0-alpha.6 [UNRELEASED]
+
+## 3.0.0-alpha.5 [2022-10-10]
+
+- Bumps `fvm_ipld_encoding` and switches from `cs_serde_bytes` to `fvm_ipld_encoding::strict_bytes`.
+
+## 3.0.0-alpha.4 [2022-10-10]
+
+- Small f4 address fixes.
+
+## 3.0.0-alpha.3 [2022-10-10]
+
+- Switch to rust 2021 edition.
+- Add network version 18.
+- BREAKING: Allow changing the address "network" at runtime.
+- BREAKING: Update the f4 address format and include a checksum.
+- BREAKING: Add the gas premium and gas limit to the `vm::context` return type.
 
 ## 3.0.0-alpha.2 [2022-09-16]